Maya Rule of Thirds and Fives
This is an overlay for any camera in maya but is coded to fit a cinematic camera. It will create a rule a thirds rig and hide it if pressed again, then if you remove the rule of thirds and press the fives tool it will create an overlay of the fives grid over the camera, the focal length and aspet ration will not affect the grids and can adapt.