Add files via upload
The turntable for blender - a note to add is that you will notice it spawns in a sphere and then deletes the sphere. This sphere is created to the size of the model so that the camera will centre to perfectly fit the model inframe instead of having cutt off edges when it turns.